Disassembled UniDDS HEX files.
We disassembled the original HEX files for the UniDDS-VFO
(as provided by the QRP Project, Nikolai Zenker, DL7NIK).
Then we also resynthesized them by sending them through
'gpasm' again, to ensure that we're able to recreate the
original HEX files from scratch.
Like this, we have a complete toolchain for the development
of our new "HW9-DDS-VFO" firmware now. ;)